requirements.txt
adafruit-io==2.5.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.4
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astor==0.8.1
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
blessed==1.20.0
blinker==1.6.2
cachelib==0.10.2
cachetools==4.2.4
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
cryptography==40.0.2
cycler==0.11.0
dataclasses-json==0.5.7
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
docopt==0.6.2
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.8.0
firebase==3.0.1
Flask==2.3.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.9.0
gcloud==0.18.3
git-python==1.0.3
gitdb==4.0.10
GitPython==3.1.36
google-api-core==1.24.1
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-cloud==0.34.0
googleapis-common-protos==1.52.0
greenlet==2.0.2
gspread==3.6.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
inquirer==3.1.3
ipykernel==6.23.1
#ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.4.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
jws==0.1.3
kiwisolver==1.4.5
langchain==0.0.167
litellm==0.1.609
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.3
notebook_shim==0.2.2
numexpr==2.8.4
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.1.0
open-interpreter==0.1.3
# openai==0.27.6
openapi-schema-pydantic==1.2.4
outcome==1.2.0
packaging==23.1
paho-mqtt==1.6.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pipreqs==0.4.13
# platformdirs==3.5.1
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==1.10.7
Pygments==2.15.1
PyJWT==2.0.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
Pyrebase4==4.7.1
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-json-logger==2.0.7
python-jwt==2.0.1
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.1
QtPy==2.3.0
readchar==4.0.5
regex==2023.5.5
requests==2.29.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rsa==4.9
seaborn==0.12.2
selenium==4.2.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SQLAlchemy==2.0.13
stack-data==0.6.2
tenacity==8.2.2
terminado==0.17.1
tiktoken==0.4.0
tinycss2==1.2.1
tokentrim==0.1.10
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
twilio==6.50.1
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.15
urllib3-secure-extra==0.1.0
virtualenv==20.16.6
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.3
wget==3.2
widgetsnbextension==4.0.5
wrapt==1.12.1
wsproto==1.2.0
WTForms==3.0.1
xlwt==1.3.0
yarg==0.1.9
yarl==1.9.2
zipp==3.16.2
